DSS: MANAGER: Fix buffer overflow.
Zoom3 does not have a third manager.
This buffer over-flow was causing a corruption
on the isr function pointer array,
on the disply irq.

The loop in disply irq checked for null
on the isr to decide if the isr should be called.
because of the corruption, the isr array value
would be any random value, which would be interpreted
as a function pointer. When the function pointer
got called pointing to garbage, A kernel panic
would follow.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>